python-pipeline/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
numpy==1.26.4
python-dotenv==1.0.1
